refactors aes-siv, removing dynamic ctx allocation
