+++ urcrypt/aes_siv.c
@@ -2,38 +2,34 @@
 #include "util.h"
 #include <aes_siv.h>
 
-static AES_SIV_CTX*
-urcrypt__aes_siv_init(uint8_t *key,
+/* the caller stack-allocates the context (see aes_siv.h); urcrypt never calls
+ * malloc, and we wipe the finished context with AES_SIV_CTX_cleanup().
+ */
+static int
+urcrypt__aes_siv_init(AES_SIV_CTX *ctx,
+                      uint8_t *key,
                       size_t key_length,
                       urcrypt_aes_siv_data *data,
                       size_t data_length)
 {
-  AES_SIV_CTX *ctx = AES_SIV_CTX_new();
-  if ( NULL == ctx ) {
-    return NULL;
+  urcrypt__reverse(key_length, key);
+  if ( 0 == AES_SIV_Init(ctx, key, key_length) ) {
+    return -1;
   }
   else {
-    urcrypt__reverse(key_length, key);
-    if ( 0 == AES_SIV_Init(ctx, key, key_length) ) {
-      AES_SIV_CTX_free(ctx);
-      return NULL;
-    }
-    else {
-      size_t i, len;
-      uint8_t *dat;
+    size_t i, len;
+    uint8_t *dat;
 
-      for ( i = 0; i < data_length; ++i ) {
-        len = data[i].length;
-        dat = data[i].bytes;
-        urcrypt__reverse(len, dat);
-        if ( 0 == AES_SIV_AssociateData(ctx, dat, len) ) {
-          AES_SIV_CTX_free(ctx);
-          return NULL;
-        }
+    for ( i = 0; i < data_length; ++i ) {
+      len = data[i].length;
+      dat = data[i].bytes;
+      urcrypt__reverse(len, dat);
+      if ( 0 == AES_SIV_AssociateData(ctx, dat, len) ) {
+        return -1;
       }
-
-      return ctx;
     }
+
+    return 0;
   }
 }
 
@@ -47,26 +43,26 @@ urcrypt__aes_siv_en(uint8_t *key,
                     uint8_t iv[16],
                     uint8_t *out)
 {
-  AES_SIV_CTX *ctx = urcrypt__aes_siv_init(key, key_length, data, data_length);
+  int ret;
+  AES_SIV_CTX ctx;
 
-  if ( NULL == ctx ) {
-    return -1;
+  if ( 0 != urcrypt__aes_siv_init(&ctx, key, key_length, data, data_length) ) {
+    ret = -1;
   }
   else {
-    int ret;
     urcrypt__reverse(message_length, message);
-    ret = AES_SIV_EncryptFinal(ctx, iv, out, message, message_length);
-    AES_SIV_CTX_free(ctx);
-
-    if ( 0 == ret ) {
-      return -2;
+    if ( 0 == AES_SIV_EncryptFinal(&ctx, iv, out, message, message_length) ) {
+      ret = -2;
     }
     else {
       urcrypt__reverse(16, iv);
       urcrypt__reverse(message_length, out);
-      return 0;
+      ret = 0;
     }
   }
+
+  AES_SIV_CTX_cleanup(&ctx);
+  return ret;
 }
 
 static int
@@ -79,27 +75,26 @@ urcrypt__aes_siv_de(uint8_t *key,
                     uint8_t iv[16],
                     uint8_t *out)
 {
-  AES_SIV_CTX *ctx = urcrypt__aes_siv_init(key, key_length, data, data_length);
+  int ret;
+  AES_SIV_CTX ctx;
 
-  if ( NULL == ctx ) {
-    return -1;
+  if ( 0 != urcrypt__aes_siv_init(&ctx, key, key_length, data, data_length) ) {
+    ret = -1;
   }
   else {
-    int ret;
-
     urcrypt__reverse(message_length, message);
     urcrypt__reverse(16, iv);
-    ret = AES_SIV_DecryptFinal(ctx, out, iv, message, message_length);
-    AES_SIV_CTX_free(ctx);
-
-    if ( 0 == ret ) {
-      return -2;
+    if ( 0 == AES_SIV_DecryptFinal(&ctx, out, iv, message, message_length) ) {
+      ret = -2;
     }
     else {
       urcrypt__reverse(message_length, out);
-      return 0;
+      ret = 0;
     }
   }
+
+  AES_SIV_CTX_cleanup(&ctx);
+  return ret;
 }
 
 int
